requirements.txt
alembic==1.0.8
amqp==2.4.2
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.2.1
attrs==18.2.0
Authlib==0.10
Automat==0.6.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.6.0.0
bleach==2.1.3
celery==4.3.0
centrifuge==0.8.4
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.7
configparser==3.5.0
constantly==15.1.0
coverage==4.5.2
cryptography==2.4.1
cycler==0.10.0
decorator==4.4.0
dnspython==1.15.0
entrypoints==0.2.3
enum34==1.1.6
Flask==1.0.2
Flask-Cors==3.0.7
Flask-JSONRPC==0.3.1
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gunicorn==19.9.0
hiredis==0.3.0
html5lib==1.0.1
hyperlink==17.3.1
idna==2.7
incremental==16.10.1
infinity==1.4
intervals==0.8.1
ipaddress==1.0.22
ipykernel==4.8.2
ipython==5.6.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
kombu==4.5.0
lazy-object-proxy==1.3.1
logilab-common==1.4.1
Mako==1.0.8
Markdown==2.6.9
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.3.0
mpmath==1.0.0
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.4.1
numpy==1.14.3
PAM==0.4.2
pandas==0.22.0
pandocfilters==1.4.2
pathlib2==2.3.2
pexpect==4.5.0
pickleshare==0.7.4
pluggy==0.7.1
prompt-toolkit==1.0.15
psycopg2==2.7.5
ptyprocess==0.5.2
py==1.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pyflakes==1.6.0
pygame===1.9.1release
Pygments==2.2.0
pygobject==3.26.1
pylint==1.9.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyserial==3.4
pytest==3.8.1
pytest-flask==0.12.0
python-dateutil==2.8.0
python-editor==1.0.4
python-libtorrent==1.1.5
python-memcached==1.59
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.20.1
scandir==1.9.0
scipy==1.1.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==16.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
sockjs-tornado==1.0.1
SQLAlchemy==1.3.2
SQLAlchemy-Utils==0.33.11
subprocess32==3.5.0
sympy==1.1.1
terminado==0.8.1
testpath==0.3.1
toredis-fork==0.1.4
tornado==5.0.2
toro==0.8
traitlets==4.3.2
ttystatus==0.38
Twisted==17.9.0
unity-lens-photos==1.0
urllib3==1.24.1
validators==0.12.4
vine==1.3.0
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.2.1
wrapt==1.10.11
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
zope.interface==4.3.2
